Cleaned up tests a bit
