+++ tests/V5/tests.hs
@@ -18,51 +18,30 @@
 #endif
 module Main where
 
-import           Control.Applicative
+import           Test.Import
+
 import           Control.Error                   hiding (Script)
-import           Control.Exception               (evaluate)
-import           Control.Monad
-import           Control.Monad.Catch
-import           Control.Monad.Reader
-import           Data.Aeson
-import           Data.Aeson.TH
-import           Data.Aeson.Types                (parseEither)
-import qualified Data.ByteString.Lazy.Char8      as BL8
 import qualified Data.HashMap.Strict             as HM
 import           Data.List                       (nub)
 import qualified Data.List                       as L
 import           Data.List.NonEmpty              (NonEmpty (..))
 import qualified Data.List.NonEmpty              as NE
 import qualified Data.Map.Strict                 as M
-import           Data.Monoid
-import           Data.Ord                        (comparing)
-import           Data.Proxy
 import           Data.Text                       (Text)
 import qualified Data.Text                       as T
 import           Data.Time.Calendar              (Day (..), fromGregorian)
-import           Data.Time.Clock                 (NominalDiffTime, UTCTime (..),
+import           Data.Time.Clock                 (UTCTime (..),
                                                   secondsToDiffTime)
-import           Data.Typeable
 import qualified Data.Vector                     as V
 import qualified Data.Version                    as Vers
-import           Database.V5.Bloodhound
-import           Lens.Micro
-import           Lens.Micro.Aeson
+import qualified Database.V5.Bloodhound
 import qualified Lens.Micro.Aeson                as LMA
-import           Network.HTTP.Client             hiding (Proxy)
 import qualified Network.HTTP.Types.Method       as NHTM
 import qualified Network.HTTP.Types.Status       as NHTS
 import qualified Network.URI                     as URI
 import           Prelude                         hiding (filter)
-import           System.IO.Temp
-import           System.PosixCompat.Files
-import           Test.Hspec
-import           Test.QuickCheck.Property.Monoid (T (..), eq, prop_Monoid)
 
-import           Test.Hspec.QuickCheck           (prop)
-import           Test.QuickCheck
-import           Test.QuickCheck.TH.Generators
--- import           Text.Pretty.Simple (pPrint)
+import qualified Test.JSON as JSON
 
 testServer  :: Server
 testServer  = Server "http://localhost:9200"
@@ -155,22 +134,6 @@ is v = getServerVersion >>= \x -> return $ x == Just v
 
 when' :: Monad m => m Bool -> m () -> m ()
 when' b f = b >>= \x -> when x f
-
-propJSON :: forall a
-          . ( Arbitrary a
-            , ToJSON a
-            , FromJSON a
-            , Show a
-            , Eq a
-            , Typeable a
-            )
-         => Proxy a -> Spec
-propJSON _ = prop testName $ \(a :: a) ->
-  let jsonStr = "via " <> BL8.unpack (encode a)
-  in counterexample jsonStr (parseEither parseJSON (toJSON a)
-                             === Right a)
-  where testName = show ty <> " FromJSON/ToJSON roundtrips"
-        ty = typeOf (undefined :: a)
 
 data Location = Location { lat :: Double
                          , lon :: Double } deriving (Eq, Show)
@@ -451,114 +414,9 @@ instance FromJSON BulkTest where
         t <- o .: "name"
         BulkTest <$> parseJSON t
 
--- | Typeclass for "equal where it matters". Use this to specify
--- less-strict equivalence for things such as lists that can wind up
--- in an unpredictable order
-class ApproxEq a where
-  (=~) :: a -> a -> Bool
-
-  showApproxEq :: a -> String
-  default showApproxEq :: (Show a) => a -> String
-  showApproxEq = show
-
-(==~) :: (ApproxEq a) => a -> a -> Property
-a ==~ b = counterexample (showApproxEq a ++ " !=~ " ++ showApproxEq b) (a =~ b)
-
-propApproxJSON :: forall a
-                . ( Arbitrary a
-                  , ToJSON a
-                  , FromJSON a
-                  , Show a
-                  , ApproxEq a
-                  , Typeable a
-                  )
-               => Proxy a -> Spec
-propApproxJSON _ = prop testName $ \(a :: a) ->
-  let jsonStr = "via " <> BL8.unpack (encode a)
-  in counterexample jsonStr (parseEither parseJSON (toJSON a)
-                             ==~ Right a)
-  where testName = show ty <> " FromJSON/ToJSON roundtrips"
-        ty = typeOf (undefined :: a)
-
-instance ApproxEq NominalDiffTime where (=~) = (==)
-instance ApproxEq Bool where (=~) = (==)
-instance ApproxEq Int where (=~) = (==)
-instance (Eq a, Show a) => ApproxEq (Maybe a) where (=~) = (==)
-instance ApproxEq Char where
-  (=~) = (==)
-
-instance ApproxEq NodeAttrFilter where (=~) = (==)
-instance ApproxEq NodeAttrName where (=~) = (==)
-instance (Eq a, Show a) => ApproxEq (NonEmpty a) where (=~) = (==)
-instance (ApproxEq l, Show l, ApproxEq r, Show r) => ApproxEq (Either l r) where
-  Left a =~ Left b = a =~ b
-  Right a =~ Right b = a =~ b
-  _ =~ _ = False
-  showApproxEq (Left x)  = "Left " <> showApproxEq x
-  showApproxEq (Right x) = "Right " <> showApproxEq x
-instance (ApproxEq a, Show a) => ApproxEq [a] where
-  as =~ bs = and (zipWith (=~) as bs)
-instance ApproxEq ReplicaCount where (=~) = (==)
-instance ApproxEq ReplicaBounds where (=~) = (==)
-instance ApproxEq Bytes where (=~) = (==)
-instance ApproxEq AllocationPolicy where (=~) = (==)
-instance ApproxEq InitialShardCount where (=~) = (==)
-instance ApproxEq FSType where (=~) = (==)
-
 noDuplicates :: Eq a => [a] -> Bool
 noDuplicates xs = nub xs == xs
 
-instance Arbitrary NominalDiffTime where
-  arbitrary = fromInteger <$> arbitrary
-
-#if !MIN_VERSION_QuickCheck(2,8,0)
-instance (Arbitrary k, Ord k, Arbitrary v) => Arbitrary (M.Map k v) where
-  arbitrary = M.fromList <$> arbitrary
-#endif
-
-instance Arbitrary Text where
-  arbitrary = T.pack <$> arbitrary
-
-instance Arbitrary UTCTime where
-  arbitrary = UTCTime
-          <$> arbitrary
-          <*> (fromRational . toRational <$> choose (0::Double, 86400))
-
-instance Arbitrary Day where
-    arbitrary = ModifiedJulianDay <$> (2000 +) <$> arbitrary
-    shrink    = (ModifiedJulianDay <$>) . shrink . toModifiedJulianDay
-
-#if !MIN_VERSION_QuickCheck(2,9,0)
-instance Arbitrary a => Arbitrary (NonEmpty a) where
-  arbitrary = liftA2 (:|) arbitrary arbitrary
-#endif
-
-arbitraryScore :: Gen Score
-arbitraryScore = fmap getPositive <$> arbitrary
-
-instance (Arbitrary a, Typeable a) => Arbitrary (Hit a) where
-  arbitrary = Hit <$> arbitrary
-                  <*> arbitrary
-                  <*> arbitrary
-                  <*> arbitraryScore
-                  <*> arbitrary
-                  <*> arbitrary
-                  <*> arbitrary
-
-instance Arbitrary HitFields where
-  arbitrary = pure (HitFields M.empty)
-  shrink = const []
-
-instance (Arbitrary a, Typeable a) => Arbitrary (SearchHits a) where
-  arbitrary = reduceSize $ do
-    tot <- getPositive <$> arbitrary
-    score <- arbitraryScore
-    hs <- arbitrary
-    return $ SearchHits tot score hs
-
-
-reduceSize :: Gen a -> Gen a
-reduceSize f = sized $ \n -> resize (n `div` 2) f
 
 getSource :: EsResult a -> Maybe a
 getSource = fmap _source . foundResult
@@ -570,487 +428,9 @@ grabFirst r =
     (Right Nothing)  -> Left (EsError 500 "Source was missing")
     (Right (Just x)) -> Right x
 
--------------------------------------------------------------------------------
-arbitraryAlphaNum :: Gen Char
-arbitraryAlphaNum = oneof [choose ('a', 'z')
-                          ,choose ('A','Z')
-                          , choose ('0', '9')]
-
-instance Arbitrary RoutingValue where
-  arbitrary = RoutingValue . T.pack <$> listOf1 arbitraryAlphaNum
-
-instance Arbitrary AliasRouting where
-  arbitrary = oneof [allAlias
-                    ,one
-                    ,theOther
-                    ,both']
-    where one = GranularAliasRouting
-                <$> (Just <$> arbitrary)
-                <*> pure Nothing
-          theOther = GranularAliasRouting Nothing
-                     <$> (Just <$> arbitrary)
-          both' = GranularAliasRouting
-                 <$> (Just <$> arbitrary)
-                 <*> (Just <$> arbitrary)
-          allAlias = AllAliasRouting <$> arbitrary
-
-
-
-instance Arbitrary FieldName where
-  arbitrary =
-        FieldName
-      . T.pack
-    <$> listOf1 arbitraryAlphaNum
-
-
-#if MIN_VERSION_base(4,10,0)
--- Test.QuickCheck.Modifiers
-
-qcNonEmptyToNonEmpty :: NonEmptyList a -> NonEmpty a
-qcNonEmptyToNonEmpty (NonEmpty (a : xs)) = (a :| xs)
-qcNonEmptyToNonEmpty (NonEmpty []) = error "NonEmpty was empty!"
-
-instance Arbitrary a => Arbitrary (NonEmpty a) where
-  arbitrary =
-        qcNonEmptyToNonEmpty
-    <$> arbitrary
-#endif
-instance Arbitrary ScriptFields where
-  arbitrary =
-    pure $ ScriptFields $
-      HM.fromList []
-
-  shrink = const []
-
-instance Arbitrary ScriptParams where
-  arbitrary =
-    pure $ ScriptParams $
-      HM.fromList [ ("a", Number 42)
-                  , ("b", String "forty two")
-                  ]
-
-  shrink = const []
-
-instance Arbitrary RegexpFlags where
-  arbitrary = oneof [ pure AllRegexpFlags
-                    , pure NoRegexpFlags
-                    , SomeRegexpFlags <$> genUniqueFlags
-                    ]
-    where genUniqueFlags =
-                NE.fromList . nub
-            <$> listOf1 arbitrary
-
-instance Arbitrary IndexAliasCreate where
-  arbitrary =
-        IndexAliasCreate
-    <$> arbitrary
-    <*> reduceSize arbitrary
-
-instance Arbitrary Query where
-  arbitrary =
-    reduceSize
-    $ oneof [ TermQuery <$> arbitrary <*> arbitrary
-            , TermsQuery <$> arbitrary <*> arbitrary
-            , QueryMatchQuery <$> arbitrary
-            , QueryMultiMatchQuery <$> arbitrary
-            , QueryBoolQuery <$> arbitrary
-            , QueryBoostingQuery <$> arbitrary
-            , QueryCommonTermsQuery <$> arbitrary
-            , ConstantScoreQuery <$> arbitrary <*> arbitrary
-            , QueryDisMaxQuery <$> arbitrary
-            , QueryFuzzyLikeThisQuery <$> arbitrary
-            , QueryFuzzyLikeFieldQuery <$> arbitrary
-            , QueryFuzzyQuery <$> arbitrary
-            , QueryHasChildQuery <$> arbitrary
-            , QueryHasParentQuery <$> arbitrary
-            , IdsQuery <$> arbitrary <*> arbitrary
-            , QueryIndicesQuery <$> arbitrary
-            , MatchAllQuery <$> arbitrary
-            , QueryMoreLikeThisQuery <$> arbitrary
-            , QueryMoreLikeThisFieldQuery <$> arbitrary
-            , QueryNestedQuery <$> arbitrary
-            , QueryPrefixQuery <$> arbitrary
-            , QueryQueryStringQuery <$> arbitrary
-            , QuerySimpleQueryStringQuery <$> arbitrary
-            , QueryRangeQuery <$> arbitrary
-            , QueryRegexpQuery <$> arbitrary
-            , QueryTemplateQueryInline <$> arbitrary
-            ]
-  -- TODO: Implement shrink
-  -- shrink = genericShrink
-
-instance Arbitrary Filter where
-  arbitrary =
-    Filter <$> arbitrary 
-  shrink (Filter q) =
-    Filter <$> shrink q
-
-instance Arbitrary ReplicaBounds where
-  arbitrary = oneof [ replicasBounded
-                    , replicasLowerBounded
-                    , pure ReplicasUnbounded
-                    ]
-    where replicasBounded = do
-            Positive a <- arbitrary
-            Positive b <- arbitrary
-            return (ReplicasBounded a b)
-          replicasLowerBounded = do
-            Positive a <- arbitrary
-            return (ReplicasLowerBounded a)
-
-instance Arbitrary NodeAttrName where
-  arbitrary =
-        NodeAttrName
-      . T.pack
-    <$> listOf1 arbitraryAlphaNum
-
-
-instance Arbitrary NodeAttrFilter where
-  arbitrary = do
-    n <- arbitrary
-    s:ss <- listOf1 (listOf1 arbitraryAlphaNum)
-    let ts = T.pack <$> s :| ss
-    return (NodeAttrFilter n ts)
-
-instance Arbitrary VersionNumber where
-  arbitrary = mk . fmap getPositive . getNonEmpty <$> arbitrary
-    where
-      mk versions = VersionNumber (Vers.Version versions [])
-
-instance Arbitrary TemplateQueryKeyValuePairs where
-  arbitrary = TemplateQueryKeyValuePairs . HM.fromList <$> arbitrary
-  shrink (TemplateQueryKeyValuePairs x) = map (TemplateQueryKeyValuePairs . HM.fromList) . shrink $ HM.toList x
-
-makeArbitrary ''IndexName
-instance Arbitrary IndexName where arbitrary = arbitraryIndexName
-makeArbitrary ''MappingName
-instance Arbitrary MappingName where arbitrary = arbitraryMappingName
-makeArbitrary ''DocId
-instance Arbitrary DocId where arbitrary = arbitraryDocId
-makeArbitrary ''Version
-instance Arbitrary Version where arbitrary = arbitraryVersion
-makeArbitrary ''BuildHash
-instance Arbitrary BuildHash where arbitrary = arbitraryBuildHash
-makeArbitrary ''IndexAliasRouting
-instance Arbitrary IndexAliasRouting where arbitrary = arbitraryIndexAliasRouting
-makeArbitrary ''ShardCount
-instance Arbitrary ShardCount where arbitrary = arbitraryShardCount
-makeArbitrary ''ReplicaCount
-instance Arbitrary ReplicaCount where arbitrary = arbitraryReplicaCount
-makeArbitrary ''TemplateName
-instance Arbitrary TemplateName where arbitrary = arbitraryTemplateName
-makeArbitrary ''TemplatePattern
-instance Arbitrary TemplatePattern where arbitrary = arbitraryTemplatePattern
-makeArbitrary ''QueryString
-instance Arbitrary QueryString where arbitrary = arbitraryQueryString
-makeArbitrary ''CacheName
-instance Arbitrary CacheName where arbitrary = arbitraryCacheName
-makeArbitrary ''CacheKey
-instance Arbitrary CacheKey where arbitrary = arbitraryCacheKey
-makeArbitrary ''Existence
-instance Arbitrary Existence where arbitrary = arbitraryExistence
-makeArbitrary ''CutoffFrequency
-instance Arbitrary CutoffFrequency where arbitrary = arbitraryCutoffFrequency
-makeArbitrary ''Analyzer
-instance Arbitrary Analyzer where arbitrary = arbitraryAnalyzer
-makeArbitrary ''MaxExpansions
-instance Arbitrary MaxExpansions where arbitrary = arbitraryMaxExpansions
-makeArbitrary ''Lenient
-instance Arbitrary Lenient where arbitrary = arbitraryLenient
-makeArbitrary ''Tiebreaker
-instance Arbitrary Tiebreaker where arbitrary = arbitraryTiebreaker
-makeArbitrary ''Boost
-instance Arbitrary Boost where arbitrary = arbitraryBoost
-makeArbitrary ''BoostTerms
-instance Arbitrary BoostTerms where arbitrary = arbitraryBoostTerms
-makeArbitrary ''MinimumMatch
-instance Arbitrary MinimumMatch where arbitrary = arbitraryMinimumMatch
-makeArbitrary ''DisableCoord
-instance Arbitrary DisableCoord where arbitrary = arbitraryDisableCoord
-makeArbitrary ''IgnoreTermFrequency
-instance Arbitrary IgnoreTermFrequency where arbitrary = arbitraryIgnoreTermFrequency
-makeArbitrary ''MinimumTermFrequency
-instance Arbitrary MinimumTermFrequency where arbitrary = arbitraryMinimumTermFrequency
-makeArbitrary ''MaxQueryTerms
-instance Arbitrary MaxQueryTerms where arbitrary = arbitraryMaxQueryTerms
-makeArbitrary ''Fuzziness
-instance Arbitrary Fuzziness where arbitrary = arbitraryFuzziness
-makeArbitrary ''PrefixLength
-instance Arbitrary PrefixLength where arbitrary = arbitraryPrefixLength
-makeArbitrary ''TypeName
-instance Arbitrary TypeName where arbitrary = arbitraryTypeName
-makeArbitrary ''PercentMatch
-instance Arbitrary PercentMatch where arbitrary = arbitraryPercentMatch
-makeArbitrary ''StopWord
-instance Arbitrary StopWord where arbitrary = arbitraryStopWord
-makeArbitrary ''QueryPath
-instance Arbitrary QueryPath where arbitrary = arbitraryQueryPath
-makeArbitrary ''AllowLeadingWildcard
-instance Arbitrary AllowLeadingWildcard where arbitrary = arbitraryAllowLeadingWildcard
-makeArbitrary ''LowercaseExpanded
-instance Arbitrary LowercaseExpanded where arbitrary = arbitraryLowercaseExpanded
-makeArbitrary ''EnablePositionIncrements
-instance Arbitrary EnablePositionIncrements where arbitrary = arbitraryEnablePositionIncrements
-makeArbitrary ''AnalyzeWildcard
-instance Arbitrary AnalyzeWildcard where arbitrary = arbitraryAnalyzeWildcard
-makeArbitrary ''GeneratePhraseQueries
-instance Arbitrary GeneratePhraseQueries where arbitrary = arbitraryGeneratePhraseQueries
-makeArbitrary ''Locale
-instance Arbitrary Locale where arbitrary = arbitraryLocale
-makeArbitrary ''MaxWordLength
-instance Arbitrary MaxWordLength where arbitrary = arbitraryMaxWordLength
-makeArbitrary ''MinWordLength
-instance Arbitrary MinWordLength where arbitrary = arbitraryMinWordLength
-makeArbitrary ''PhraseSlop
-instance Arbitrary PhraseSlop where arbitrary = arbitraryPhraseSlop
-makeArbitrary ''MinDocFrequency
-instance Arbitrary MinDocFrequency where arbitrary = arbitraryMinDocFrequency
-makeArbitrary ''MaxDocFrequency
-instance Arbitrary MaxDocFrequency where arbitrary = arbitraryMaxDocFrequency
-makeArbitrary ''Regexp
-instance Arbitrary Regexp where arbitrary = arbitraryRegexp
-makeArbitrary ''SimpleQueryStringQuery
-instance Arbitrary SimpleQueryStringQuery where arbitrary = arbitrarySimpleQueryStringQuery
-makeArbitrary ''FieldOrFields
-instance Arbitrary FieldOrFields where arbitrary = arbitraryFieldOrFields
-makeArbitrary ''SimpleQueryFlag
-instance Arbitrary SimpleQueryFlag where arbitrary = arbitrarySimpleQueryFlag
-makeArbitrary ''RegexpQuery
-instance Arbitrary RegexpQuery where arbitrary = arbitraryRegexpQuery
-makeArbitrary ''QueryStringQuery
-instance Arbitrary QueryStringQuery where arbitrary = arbitraryQueryStringQuery
-makeArbitrary ''RangeQuery
-instance Arbitrary RangeQuery where arbitrary = arbitraryRangeQuery
-makeArbitrary ''RangeValue
-instance Arbitrary RangeValue where arbitrary = arbitraryRangeValue
-makeArbitrary ''PrefixQuery
-instance Arbitrary PrefixQuery where arbitrary = arbitraryPrefixQuery
-makeArbitrary ''NestedQuery
-instance Arbitrary NestedQuery where arbitrary = arbitraryNestedQuery
-makeArbitrary ''MoreLikeThisFieldQuery
-instance Arbitrary MoreLikeThisFieldQuery where arbitrary = arbitraryMoreLikeThisFieldQuery
-makeArbitrary ''MoreLikeThisQuery
-instance Arbitrary MoreLikeThisQuery where arbitrary = arbitraryMoreLikeThisQuery
-makeArbitrary ''IndicesQuery
-instance Arbitrary IndicesQuery where arbitrary = arbitraryIndicesQuery
-makeArbitrary ''HasParentQuery
-instance Arbitrary HasParentQuery where arbitrary = arbitraryHasParentQuery
-makeArbitrary ''HasChildQuery
-instance Arbitrary HasChildQuery where arbitrary = arbitraryHasChildQuery
-makeArbitrary ''FuzzyQuery
-instance Arbitrary FuzzyQuery where arbitrary = arbitraryFuzzyQuery
-makeArbitrary ''FuzzyLikeFieldQuery
-instance Arbitrary FuzzyLikeFieldQuery where arbitrary = arbitraryFuzzyLikeFieldQuery
-makeArbitrary ''FuzzyLikeThisQuery
-instance Arbitrary FuzzyLikeThisQuery where arbitrary = arbitraryFuzzyLikeThisQuery
-makeArbitrary ''DisMaxQuery
-instance Arbitrary DisMaxQuery where arbitrary = arbitraryDisMaxQuery
-makeArbitrary ''CommonTermsQuery
-instance Arbitrary CommonTermsQuery where arbitrary = arbitraryCommonTermsQuery
-makeArbitrary ''DistanceRange
-instance Arbitrary DistanceRange where arbitrary = arbitraryDistanceRange
-makeArbitrary ''MultiMatchQuery
-instance Arbitrary MultiMatchQuery where arbitrary = arbitraryMultiMatchQuery
-makeArbitrary ''LessThanD
-instance Arbitrary LessThanD where arbitrary = arbitraryLessThanD
-makeArbitrary ''LessThanEqD
-instance Arbitrary LessThanEqD where arbitrary = arbitraryLessThanEqD
-makeArbitrary ''GreaterThanD
-instance Arbitrary GreaterThanD where arbitrary = arbitraryGreaterThanD
-makeArbitrary ''GreaterThanEqD
-instance Arbitrary GreaterThanEqD where arbitrary = arbitraryGreaterThanEqD
-makeArbitrary ''LessThan
-instance Arbitrary LessThan where arbitrary = arbitraryLessThan
-makeArbitrary ''LessThanEq
-instance Arbitrary LessThanEq where arbitrary = arbitraryLessThanEq
-makeArbitrary ''GreaterThan
-instance Arbitrary GreaterThan where arbitrary = arbitraryGreaterThan
-makeArbitrary ''GreaterThanEq
-instance Arbitrary GreaterThanEq where arbitrary = arbitraryGreaterThanEq
-makeArbitrary ''GeoPoint
-instance Arbitrary GeoPoint where arbitrary = arbitraryGeoPoint
-makeArbitrary ''NullValue
-instance Arbitrary NullValue where arbitrary = arbitraryNullValue
-makeArbitrary ''MinimumMatchHighLow
-instance Arbitrary MinimumMatchHighLow where arbitrary = arbitraryMinimumMatchHighLow
-makeArbitrary ''CommonMinimumMatch
-instance Arbitrary CommonMinimumMatch where arbitrary = arbitraryCommonMinimumMatch
-makeArbitrary ''BoostingQuery
-instance Arbitrary BoostingQuery where arbitrary = arbitraryBoostingQuery
-makeArbitrary ''BoolQuery
-instance Arbitrary BoolQuery where arbitrary = arbitraryBoolQuery
-makeArbitrary ''MatchQuery
-instance Arbitrary MatchQuery where arbitrary = arbitraryMatchQuery
-makeArbitrary ''MultiMatchQueryType
-instance Arbitrary MultiMatchQueryType where arbitrary = arbitraryMultiMatchQueryType
-makeArbitrary ''BooleanOperator
-instance Arbitrary BooleanOperator where arbitrary = arbitraryBooleanOperator
-makeArbitrary ''ZeroTermsQuery
-instance Arbitrary ZeroTermsQuery where arbitrary = arbitraryZeroTermsQuery
-makeArbitrary ''MatchQueryType
-instance Arbitrary MatchQueryType where arbitrary = arbitraryMatchQueryType
-makeArbitrary ''SearchAliasRouting
-instance Arbitrary SearchAliasRouting where arbitrary = arbitrarySearchAliasRouting
-makeArbitrary ''ScoreType
-instance Arbitrary ScoreType where arbitrary = arbitraryScoreType
-makeArbitrary ''Distance
-instance Arbitrary Distance where arbitrary = arbitraryDistance
-makeArbitrary ''DistanceUnit
-instance Arbitrary DistanceUnit where arbitrary = arbitraryDistanceUnit
-makeArbitrary ''DistanceType
-instance Arbitrary DistanceType where arbitrary = arbitraryDistanceType
-makeArbitrary ''OptimizeBbox
-instance Arbitrary OptimizeBbox where arbitrary = arbitraryOptimizeBbox
-makeArbitrary ''GeoBoundingBoxConstraint
-instance Arbitrary GeoBoundingBoxConstraint where arbitrary = arbitraryGeoBoundingBoxConstraint
-makeArbitrary ''GeoFilterType
-instance Arbitrary GeoFilterType where arbitrary = arbitraryGeoFilterType
-makeArbitrary ''GeoBoundingBox
-instance Arbitrary GeoBoundingBox where arbitrary = arbitraryGeoBoundingBox
-makeArbitrary ''LatLon
-instance Arbitrary LatLon where arbitrary = arbitraryLatLon
-makeArbitrary ''RangeExecution
-instance Arbitrary RangeExecution where arbitrary = arbitraryRangeExecution
-makeArbitrary ''RegexpFlag
-instance Arbitrary RegexpFlag where arbitrary = arbitraryRegexpFlag
-makeArbitrary ''BoolMatch
-instance Arbitrary BoolMatch where arbitrary = arbitraryBoolMatch
-makeArbitrary ''Term
-instance Arbitrary Term where arbitrary = arbitraryTerm
-makeArbitrary ''IndexSettings
-instance Arbitrary IndexSettings where arbitrary = arbitraryIndexSettings
-makeArbitrary ''TokenChar
-instance Arbitrary TokenChar where arbitrary = arbitraryTokenChar
-makeArbitrary ''Ngram
-instance Arbitrary Ngram where arbitrary = arbitraryNgram
-makeArbitrary ''TokenizerDefinition
-instance Arbitrary TokenizerDefinition where arbitrary = arbitraryTokenizerDefinition
-makeArbitrary ''TokenFilter
-instance Arbitrary TokenFilter where arbitrary = arbitraryTokenFilter
-makeArbitrary ''TokenFilterDefinition
-instance Arbitrary TokenFilterDefinition where arbitrary = arbitraryTokenFilterDefinition
-makeArbitrary ''Language
-instance Arbitrary Language where arbitrary = arbitraryLanguage
-makeArbitrary ''Shingle
-instance Arbitrary Shingle where arbitrary = arbitraryShingle
-makeArbitrary ''AnalyzerDefinition
-instance Arbitrary AnalyzerDefinition where arbitrary = arbitraryAnalyzerDefinition
-makeArbitrary ''Analysis
-instance Arbitrary Analysis where arbitrary = arbitraryAnalysis
-makeArbitrary ''Tokenizer
-instance Arbitrary Tokenizer where arbitrary = arbitraryTokenizer
-makeArbitrary ''UpdatableIndexSetting
-instance Arbitrary UpdatableIndexSetting where
-  arbitrary = arbitraryUpdatableIndexSetting
-makeArbitrary ''Compression
-instance Arbitrary Compression where arbitrary = arbitraryCompression
-makeArbitrary ''Bytes
-instance Arbitrary Bytes where arbitrary = arbitraryBytes
-makeArbitrary ''AllocationPolicy
-instance Arbitrary AllocationPolicy where arbitrary = arbitraryAllocationPolicy
-makeArbitrary ''InitialShardCount
-instance Arbitrary InitialShardCount where arbitrary = arbitraryInitialShardCount
-makeArbitrary ''FSType
-instance Arbitrary FSType where arbitrary = arbitraryFSType
-makeArbitrary ''CompoundFormat
-instance Arbitrary CompoundFormat where arbitrary = arbitraryCompoundFormat
-makeArbitrary ''FsSnapshotRepo
-instance Arbitrary FsSnapshotRepo where arbitrary = arbitraryFsSnapshotRepo
-makeArbitrary ''SnapshotRepoName
-instance Arbitrary SnapshotRepoName where arbitrary = arbitrarySnapshotRepoName
-makeArbitrary ''TemplateQueryInline
-instance Arbitrary TemplateQueryInline where arbitrary = arbitraryTemplateQueryInline
-makeArbitrary ''DirectGeneratorSuggestModeTypes
-instance Arbitrary DirectGeneratorSuggestModeTypes where arbitrary = arbitraryDirectGeneratorSuggestModeTypes
-makeArbitrary ''DirectGenerators
-instance Arbitrary DirectGenerators where arbitrary = arbitraryDirectGenerators
-makeArbitrary ''PhraseSuggesterCollate
-instance Arbitrary PhraseSuggesterCollate where arbitrary = arbitraryPhraseSuggesterCollate
-makeArbitrary ''PhraseSuggesterHighlighter
-instance Arbitrary PhraseSuggesterHighlighter where arbitrary = arbitraryPhraseSuggesterHighlighter
-makeArbitrary ''Size
-instance Arbitrary Size where arbitrary = arbitrarySize
-makeArbitrary ''PhraseSuggester
-instance Arbitrary PhraseSuggester where arbitrary = arbitraryPhraseSuggester
-makeArbitrary ''SuggestType
-instance Arbitrary SuggestType where arbitrary = arbitrarySuggestType
-makeArbitrary ''Suggest
-instance Arbitrary Suggest where arbitrary = arbitrarySuggest
-
-makeArbitrary ''FunctionScoreQuery
-instance Arbitrary FunctionScoreQuery where arbitrary = arbitraryFunctionScoreQuery
-
-makeArbitrary ''FunctionScoreFunction
-instance Arbitrary FunctionScoreFunction where arbitrary = arbitraryFunctionScoreFunction
-makeArbitrary ''FunctionScoreFunctions
-instance Arbitrary FunctionScoreFunctions where arbitrary = arbitraryFunctionScoreFunctions
-makeArbitrary ''ComponentFunctionScoreFunction
-instance Arbitrary ComponentFunctionScoreFunction where arbitrary = arbitraryComponentFunctionScoreFunction
-makeArbitrary ''Script
-instance Arbitrary Script where arbitrary = arbitraryScript
-makeArbitrary ''ScriptLanguage
-instance Arbitrary ScriptLanguage where arbitrary = arbitraryScriptLanguage
-makeArbitrary ''ScriptInline
-instance Arbitrary ScriptInline where arbitrary = arbitraryScriptInline
-makeArbitrary ''ScriptId
-instance Arbitrary ScriptId where arbitrary = arbitraryScriptId
-makeArbitrary ''ScoreMode
-instance Arbitrary ScoreMode where arbitrary = arbitraryScoreMode
-makeArbitrary ''BoostMode
-instance Arbitrary BoostMode where arbitrary = arbitraryBoostMode
-makeArbitrary ''Seed
-instance Arbitrary Seed where arbitrary = arbitrarySeed
-makeArbitrary ''FieldValueFactor
-instance Arbitrary FieldValueFactor where arbitrary = arbitraryFieldValueFactor
-makeArbitrary ''Weight
-instance Arbitrary Weight where arbitrary = arbitraryWeight
-makeArbitrary ''Factor
-instance Arbitrary Factor where arbitrary = arbitraryFactor
-makeArbitrary ''FactorMissingFieldValue
-instance Arbitrary FactorMissingFieldValue where arbitrary = arbitraryFactorMissingFieldValue
-makeArbitrary ''FactorModifier
-instance Arbitrary FactorModifier where arbitrary = arbitraryFactorModifier
-
-newtype UpdatableIndexSetting' =
-  UpdatableIndexSetting' UpdatableIndexSetting
-  deriving (Show, Eq, ToJSON, FromJSON, ApproxEq, Typeable)
-
-instance Arbitrary UpdatableIndexSetting' where
-  arbitrary = do
-    settings <- arbitrary
-    return $ UpdatableIndexSetting' $ case settings of
-      RoutingAllocationInclude xs ->
-        RoutingAllocationInclude (dropDuplicateAttrNames xs)
-      RoutingAllocationExclude xs ->
-        RoutingAllocationExclude (dropDuplicateAttrNames xs)
-      RoutingAllocationRequire xs ->
-        RoutingAllocationRequire (dropDuplicateAttrNames xs)
-      x -> x
-    where
-      dropDuplicateAttrNames =
-        NE.fromList . L.nubBy sameAttrName . NE.toList
-      sameAttrName a b =
-        nodeAttrFilterName a == nodeAttrFilterName b
-  -- shrink (UpdatableIndexSetting' x) = map UpdatableIndexSetting' (shrink x)
-
--- | Due to the way nodeattrfilters get serialized here, they may come
--- out in a different order, but they are morally equivalent
-instance ApproxEq UpdatableIndexSetting where
-  RoutingAllocationInclude a =~ RoutingAllocationInclude b =
-    NE.sort a =~ NE.sort b
-  RoutingAllocationExclude a =~ RoutingAllocationExclude b =
-    NE.sort a =~ NE.sort b
-  RoutingAllocationRequire a =~ RoutingAllocationRequire b =
-    NE.sort a =~ NE.sort b
-  a =~ b = a == b
-  showApproxEq (RoutingAllocationInclude xs) = show (RoutingAllocationInclude (NE.sort xs))
-  showApproxEq (RoutingAllocationExclude xs) = show (RoutingAllocationExclude (NE.sort xs))
-  showApproxEq (RoutingAllocationRequire xs) = show (RoutingAllocationRequire (NE.sort xs))
-  showApproxEq x = show x
-
 main :: IO ()
 main = hspec $ do
-
+  JSON.spec
   describe "index create/delete API" $ do
     it "creates and then deletes the requested index" $ withTestEnv $ do
       -- priming state.
@@ -1881,124 +1261,3 @@ main = hspec $ do
           let Just results =
                 hitFields (head (hits (searchHits sr)))
           liftIO $ results `shouldBe` (HitFields (M.fromList [("test1", [Number 20000.0])]))
-
-  describe "Exact isomorphism JSON instances" $ do
-    propJSON (Proxy :: Proxy Version)
-    propJSON (Proxy :: Proxy IndexName)
-    propJSON (Proxy :: Proxy MappingName)
-    propJSON (Proxy :: Proxy DocId)
-    propJSON (Proxy :: Proxy IndexAliasRouting)
-    propJSON (Proxy :: Proxy RoutingValue)
-    propJSON (Proxy :: Proxy ShardCount)
-    propJSON (Proxy :: Proxy ReplicaCount)
-    propJSON (Proxy :: Proxy TemplateName)
-    propJSON (Proxy :: Proxy TemplatePattern)
-    propJSON (Proxy :: Proxy QueryString)
-    propJSON (Proxy :: Proxy FieldName)
-    propJSON (Proxy :: Proxy Script)
-    propJSON (Proxy :: Proxy ScriptLanguage)
-    propJSON (Proxy :: Proxy ScriptInline)
-    propJSON (Proxy :: Proxy ScriptId)
-    propJSON (Proxy :: Proxy ScriptParams)
-    propJSON (Proxy :: Proxy CacheName)
-    propJSON (Proxy :: Proxy CacheKey)
-    propJSON (Proxy :: Proxy Existence)
-    propJSON (Proxy :: Proxy CutoffFrequency)
-    propJSON (Proxy :: Proxy Analyzer)
-    propJSON (Proxy :: Proxy MaxExpansions)
-    propJSON (Proxy :: Proxy Lenient)
-    propJSON (Proxy :: Proxy Tiebreaker)
-    propJSON (Proxy :: Proxy Boost)
-    propJSON (Proxy :: Proxy BoostTerms)
-    propJSON (Proxy :: Proxy MinimumMatch)
-    propJSON (Proxy :: Proxy DisableCoord)
-    propJSON (Proxy :: Proxy IgnoreTermFrequency)
-    propJSON (Proxy :: Proxy MinimumTermFrequency)
-    propJSON (Proxy :: Proxy MaxQueryTerms)
-    propJSON (Proxy :: Proxy Fuzziness)
-    propJSON (Proxy :: Proxy PrefixLength)
-    propJSON (Proxy :: Proxy TypeName)
-    propJSON (Proxy :: Proxy PercentMatch)
-    propJSON (Proxy :: Proxy StopWord)
-    propJSON (Proxy :: Proxy QueryPath)
-    propJSON (Proxy :: Proxy AllowLeadingWildcard)
-    propJSON (Proxy :: Proxy LowercaseExpanded)
-    propJSON (Proxy :: Proxy EnablePositionIncrements)
-    propJSON (Proxy :: Proxy AnalyzeWildcard)
-    propJSON (Proxy :: Proxy GeneratePhraseQueries)
-    propJSON (Proxy :: Proxy Locale)
-    propJSON (Proxy :: Proxy MaxWordLength)
-    propJSON (Proxy :: Proxy MinWordLength)
-    propJSON (Proxy :: Proxy PhraseSlop)
-    propJSON (Proxy :: Proxy MinDocFrequency)
-    propJSON (Proxy :: Proxy MaxDocFrequency)
-    propJSON (Proxy :: Proxy Filter)
-    propJSON (Proxy :: Proxy Query)
-    propJSON (Proxy :: Proxy SimpleQueryStringQuery)
-    propJSON (Proxy :: Proxy FieldOrFields)
-    propJSON (Proxy :: Proxy SimpleQueryFlag)
-    propJSON (Proxy :: Proxy RegexpQuery)
-    propJSON (Proxy :: Proxy QueryStringQuery)
-    propJSON (Proxy :: Proxy RangeQuery)
-    propJSON (Proxy :: Proxy PrefixQuery)
-    propJSON (Proxy :: Proxy NestedQuery)
-    propJSON (Proxy :: Proxy MoreLikeThisFieldQuery)
-    propJSON (Proxy :: Proxy MoreLikeThisQuery)
-    propJSON (Proxy :: Proxy IndicesQuery)
-    propJSON (Proxy :: Proxy HasParentQuery)
-    propJSON (Proxy :: Proxy HasChildQuery)
-    propJSON (Proxy :: Proxy FuzzyQuery)
-    propJSON (Proxy :: Proxy FuzzyLikeFieldQuery)
-    propJSON (Proxy :: Proxy FuzzyLikeThisQuery)
-    propJSON (Proxy :: Proxy FunctionScoreQuery)
-    propJSON (Proxy :: Proxy BoostMode)
-    propJSON (Proxy :: Proxy ScoreMode)
-    propJSON (Proxy :: Proxy ComponentFunctionScoreFunction)
-    propJSON (Proxy :: Proxy FieldValueFactor)
-    propJSON (Proxy :: Proxy FactorModifier)
-    propJSON (Proxy :: Proxy DisMaxQuery)
-    propJSON (Proxy :: Proxy CommonTermsQuery)
-    propJSON (Proxy :: Proxy CommonMinimumMatch)
-    propJSON (Proxy :: Proxy BoostingQuery)
-    propJSON (Proxy :: Proxy BoolQuery)
-    propJSON (Proxy :: Proxy MatchQuery)
-    propJSON (Proxy :: Proxy MultiMatchQueryType)
-    propJSON (Proxy :: Proxy BooleanOperator)
-    propJSON (Proxy :: Proxy ZeroTermsQuery)
-    propJSON (Proxy :: Proxy MatchQueryType)
-    propJSON (Proxy :: Proxy AliasRouting)
-    propJSON (Proxy :: Proxy IndexAliasCreate)
-    propJSON (Proxy :: Proxy SearchAliasRouting)
-    propJSON (Proxy :: Proxy ScoreType)
-    propJSON (Proxy :: Proxy Distance)
-    propJSON (Proxy :: Proxy DistanceUnit)
-    propJSON (Proxy :: Proxy DistanceType)
-    propJSON (Proxy :: Proxy OptimizeBbox)
-    propJSON (Proxy :: Proxy GeoBoundingBoxConstraint)
-    propJSON (Proxy :: Proxy GeoFilterType)
-    propJSON (Proxy :: Proxy GeoBoundingBox)
-    propJSON (Proxy :: Proxy LatLon)
-    propJSON (Proxy :: Proxy RangeExecution)
-    prop "RegexpFlags FromJSON/ToJSON roundtrips, removing dups " $ \rfs ->
-      let expected = case rfs of
-                       SomeRegexpFlags fs -> SomeRegexpFlags (NE.fromList (nub (NE.toList fs)))
-                       x -> x
-      in parseEither parseJSON (toJSON rfs) === Right expected
-    propJSON (Proxy :: Proxy BoolMatch)
-    propJSON (Proxy :: Proxy Term)
-    propJSON (Proxy :: Proxy MultiMatchQuery)
-    propJSON (Proxy :: Proxy IndexSettings)
-    propJSON (Proxy :: Proxy CompoundFormat)
-    propJSON (Proxy :: Proxy TemplateQueryInline)
-    propJSON (Proxy :: Proxy Suggest)
-    propJSON (Proxy :: Proxy DirectGenerators)
-    propJSON (Proxy :: Proxy DirectGeneratorSuggestModeTypes)
-
-  describe "Approximate isomorphism JSON instances" $ do
-    propApproxJSON (Proxy :: Proxy UpdatableIndexSetting')
-    propApproxJSON (Proxy :: Proxy ReplicaCount)
-    propApproxJSON (Proxy :: Proxy ReplicaBounds)
-    propApproxJSON (Proxy :: Proxy Bytes)
-    propApproxJSON (Proxy :: Proxy AllocationPolicy)
-    propApproxJSON (Proxy :: Proxy InitialShardCount)
-    propApproxJSON (Proxy :: Proxy FSType)
